% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juice.R
\name{juice}
\alias{juice}
\title{In-line CSS styles}
\usage{
juice(html, options = NULL, css = NULL)
}
\arguments{
\item{html}{HTML as a string, or a file or URL}

\item{options}{A list of options to pass to juice.  See \url{https://github.com/Automattic/juice#options} for}

\item{css}{(Optional) CSS to apply to the file, as a string, file or URL.  If provided, ONLY this CSS will be inlined.  If you wish to pass \emph{additional} CSS in addition to that in \code{<style>} tags in your HTML, provide it as a string named \code{extraCss=} in the \code{options} parameter.}
}
\value{
A string of HTML with styles in-lined
}
\description{
This function an HTML file with \code{<style>} tags an in-lines the styles.
It does not currently in-line styles from external files or resources automatically,
though styles may be applied with the \code{css=} argument or \code{options=list(extraCss=)}
}
\examples{
some_html = "<style>a {font-size:30;}</style><a href='http://www.ropensci.org'>ROpenSci</a>"
juice(some_html)
# Don't apply style tags, but remove them
juice(some_html, options=list(applyStyleTags=FALSE, removeStyleTags=TRUE))
}

